requirements.txt
aiogram==3.17.0
python-dotenv==1.0.0
requests==2.32.3
sqlite3
aiohttp==3.9.5
pytz==2024.1